+++ src/calendar/index.js
@@ -53,6 +53,8 @@ class Calendar extends Component {
     displayLoadingIndicator: PropTypes.bool,
     /** Do not show days of other months in month page. Default = false */
     hideExtraDays: PropTypes.bool,
+    /** Always show six weeks on each month. Default = false */
+    showSixWeeks: PropTypes.bool,
     /** Handler which gets executed on day press. Default = undefined */
     onDayPress: PropTypes.func,
     /** Handler which gets executed on day long press. Default = undefined */
@@ -90,7 +92,15 @@ class Calendar extends Component {
     /** Apply custom disable color to selected day indexes */
     disabledDaysIndexes: PropTypes.arrayOf(PropTypes.number),
     /** Disable all touch events for disabled days. can be override with disableTouchEvent in markedDates*/
-    disableAllTouchEventsForDisabledDays: PropTypes.bool
+    disableAllTouchEventsForDisabledDays: PropTypes.bool,
+    /** Replace default month and year title with custom one. the function receive a date as parameter. */
+    renderHeader: PropTypes.any,
+    /** Enable the option to swipe between months. Default: false */
+    enableSwipeMonths: PropTypes.bool
+  };
+
+  static defaultProps = {
+    enableSwipeMonths: false
   };
 
   constructor(props) {
@@ -157,6 +167,20 @@ class Calendar extends Component {
     return (minDate && !dateutils.isGTE(date, minDate)) || (maxDate && !dateutils.isLTE(date, maxDate));
   }
 
+  getAccessibilityLabel = (state, day) => {
+    const today = XDate.locales[XDate.defaultLocale].today;
+    const formatAccessibilityLabel = XDate.locales[XDate.defaultLocale].formatAccessibilityLabel;
+    const isToday = state === 'today';
+    const markingLabel = this.getDateMarking(day);
+
+    if (formatAccessibilityLabel) {
+      return `${isToday ? today : ''} ${day.toString(formatAccessibilityLabel)} ${markingLabel}`;
+    }
+
+    return `${isToday ? 'today' : ''} ${day.toString('dddd d MMMM yyyy')} ${markingLabel}`;
+  }
+
+
   renderDay(day, id) {
     const minDate = parseDate(this.props.minDate);
     const maxDate = parseDate(this.props.maxDate);
@@ -178,7 +202,7 @@ class Calendar extends Component {
     const DayComp = this.getDayComponent();
     const date = day.getDate();
     const dateAsObject = xdateToData(day);
-    const accessibilityLabel = `${state === 'today' ? 'today' : ''} ${day.toString('dddd MMMM d')} ${this.getMarkingLabel(day)}`;
+    const accessibilityLabel = this.getAccessibilityLabel(state, day);
 
     return (
       <View style={{flex: 1, alignItems: 'center'}} key={id}>
@@ -261,6 +285,11 @@ class Calendar extends Component {
   }
 
   onSwipe = (gestureName) => {
+    const {enableSwipeMonths} = this.props;
+    if (!enableSwipeMonths) {
+      return;
+    }
+
     const {SWIPE_UP, SWIPE_DOWN, SWIPE_LEFT, SWIPE_RIGHT} = swipeDirections;
     switch (gestureName) {
     case SWIPE_UP:
@@ -312,7 +341,11 @@ class Calendar extends Component {
   }
 
   render() {
-    const days = dateutils.page(this.state.currentMonth, this.props.firstDay);
+    const {currentMonth} = this.state;
+    const {firstDay, showSixWeeks, hideExtraDays} = this.props;
+    const shouldShowSixWeeks = showSixWeeks && !hideExtraDays;
+    const days = dateutils.page(currentMonth, firstDay, shouldShowSixWeeks);
+
     const weeks = [];
     while (days.length) {
       weeks.push(this.renderWeek(days.splice(0, 7), weeks.length));
@@ -347,6 +380,7 @@ class Calendar extends Component {
             addMonth={this.addMonth}
             showIndicator={indicator}
             firstDay={this.props.firstDay}
+            showSixWeeks={this.props.showSixWeeks}
             renderArrow={this.props.renderArrow}
             monthFormat={this.props.monthFormat}
             hideDayNames={this.props.hideDayNames}
@@ -357,6 +391,7 @@ class Calendar extends Component {
             disableArrowLeft={this.props.disableArrowLeft}
             disableArrowRight={this.props.disableArrowRight}
             disabledDaysIndexes={this.props.disabledDaysIndexes}
+            renderHeader={this.props.renderHeader}
           />
           <View style={this.style.monthView}>{weeks}</View>
         </View>